created pivot posistions class
